requirements.txt

fastapi
uvicorn
python-multipart
PyMuPDF
transformers
newspaper3k
textblob
scikit-learn
spacy
nltk
